tasks/word_table_export/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-docx>=1.1.0
